File Browser better symlink handling and speed improvement (#38)
* Remove working_dir from _execute()
qdel would fail when run against an archived job when its main job was already deleted.

* Handle qdel rc=35 when ran against a finished job

* Determine if symlink is dir or file, and stop running 'ls -l /p/home/'
It can now descend into a symlinked directory, but this still does not work well if there is a broken symlink in the directory.
Changed it to run 'ls -ld' on the current directory instead of 'ls -l' on the parent. That was taking 10 seconds on the user's home directory on an HPC with 7000 users.

* Dig deep into symlinks and handle broken symlinks
This now doesn't hide all symlinks if one is broken, but it is very slow, and it doesn't show files.

* Revert "Dig deep into symlinks and handle broken symlinks"

This reverts commit 1657919d42ca7644dc94096fb18a28072f43f420.

* Change 'ls -ld' to 'ls -ldL' so it resolves symlinks itself
This code and UIT+ only care about dirs and files, so symlinks only get in the way. This will quickly show target files while ignoring broken symlinks.
Not to be confused with bad cholesterol.

* Removed duplicate code

* Initialize HpcPath by / operator

* Removing commits from a different PR

---------

Co-authored-by: sdc50 <sdc50@byu.net>