Add debug logging (#39)
* Add detailed logs for UIT+ calls

Show the command and working directory in logger.INFO, and then show the
execution time of the command, login node, return code, stdout, stderr,
and a brief stack trace in logger.DEBUG.

* Add username and fix unicode string

- Add username to first line of debug output
- Removes the b'' text around the stacktrace function call
- code cleanup

* Fix duplicate stacktrace lines

* Fix E126 continuation line over-indented for hanging indent

But not fixing "E127 continuation line over-indented for visual indent"
because I don't see the problem there.

* Improve json error handling

This used to error out on every get_file()

* Ignore qdel warning (#37)

* Remove working_dir from _execute()
qdel would fail when run against an archived job when its main job was already deleted.

* Handle qdel rc=35 when ran against a finished job

* Support qdel 'Unknown Job Id' result

* File Browser better symlink handling and speed improvement (#38)

* Remove working_dir from _execute()
qdel would fail when run against an archived job when its main job was already deleted.

* Handle qdel rc=35 when ran against a finished job

* Determine if symlink is dir or file, and stop running 'ls -l /p/home/'
It can now descend into a symlinked directory, but this still does not work well if there is a broken symlink in the directory.
Changed it to run 'ls -ld' on the current directory instead of 'ls -l' on the parent. That was taking 10 seconds on the user's home directory on an HPC with 7000 users.

* Dig deep into symlinks and handle broken symlinks
This now doesn't hide all symlinks if one is broken, but it is very slow, and it doesn't show files.

* Revert "Dig deep into symlinks and handle broken symlinks"

This reverts commit 1657919d42ca7644dc94096fb18a28072f43f420.

* Change 'ls -ld' to 'ls -ldL' so it resolves symlinks itself
This code and UIT+ only care about dirs and files, so symlinks only get in the way. This will quickly show target files while ignoring broken symlinks.
Not to be confused with bad cholesterol.

* Removed duplicate code

* Initialize HpcPath by / operator

* Removing commits from a different PR

---------

Co-authored-by: sdc50 <sdc50@byu.net>

* Retrieve debug stacktrace options from config file

- Load debug_stacktrace_allowlist from PyUIT yaml config file
- Always attempt to load the config file now, instead of only loading if
there is no UIT token or ID
- Condense the stdout and stderr code
- Change the regex to a raw string to avoid lint warnings
- Fix a typo in README.rst
- Remove quotes in tox.ini to get local flake8 to run correctly

---------

Co-authored-by: sdc50 <sdc50@byu.net>