Work around UIT+ RatchetHelper problem
The previous code would do a great job of trying a different login node if
one failed, but it failed when RatchetHelper was not running on any login
nodes. When that happens, the first call takes a while and then fails,
so the previous code would try a bunch of login nodes and fail, all
while spawning RatchetHelper. This code now retries the same login node
once if it gets a specific UIT+ error.

- Changes recursive call to a while loop
- Changes _call timeout from 20 to 25 seconds.
- Fixes a call() return value that now supports None.