Improve retries when connecting (#40)
* Change debug stacktrace filter from 'pyuit' to 'uit'

The PyUIT python package installs to a 'uit' folder, so the previous
'pyuit' filter only worked in a dev environment.

* Avoid 25 retries during connect()

- @robust() now defaults to 3 retries instead of 5
- connect() now specifies a working_dir of "." when running the ":"
command, which prevents an extra set of calls to get the $HOME variable
- connect() will bypass the @robust() decorator when making its initial
call(":") so that its own retry logic will not be multiplied by @robust()
- Fixed a sleep(1) that was not called in @robust()

* Bypass @robust() with _call instead

It was a little confusing to change @robust() decorator behavior based on
a keyword argument made to the wrapped function. This goes with a
different implementation that makes an internal _call() method that does
all the real work, which connect() now points to.

* Change from MaxRetriesError to UITError when not retrying

* Automatically enable retry_on_failure with multiple login nodes

* Avoid randomly choosing from empty set of nodes

* Add timeouts to requests.post()

These timeouts are applied to the following functions:
- call() 120 seconds for the occasional long command.
- put_file() 20 seconds to cover the 14 seconds for possibly setting up
RatchetHelper and then a few more seconds to get the transfer started. As
long as the transfer keeps sending data, requests.post will never timeout.
- get_file() 20 seconds is the same as put_file().
- list_dir() 30 seconds to cover 14 seconds for RatchetHelper and 10
seconds for looking at /p/home/ on Onyx.

Moved debug logging debug_end_time after r.json() because it wasn't
capturing the full time needed by get_file's requests.post().

* Work around UIT+ RatchetHelper problem

The previous code would do a great job of trying a different login node if
one failed, but it failed when RatchetHelper was not running on any login
nodes. When that happens, the first call takes a while and then fails,
so the previous code would try a bunch of login nodes and fail, all
while spawning RatchetHelper. This code now retries the same login node
once if it gets a specific UIT+ error.

- Changes recursive call to a while loop
- Changes _call timeout from 20 to 25 seconds.
- Fixes a call() return value that now supports None.

* Speed up file browsing with only one HPC call

Switching to a new directory in the File Browser GUI used to make two
separate UIT+ calls, one exec and one listdirectory. This change removes
the exec which was only used to determine if it is a directory or file.

* Undo some changes and let @robust handle it

I realized that @robust() was already trying to work around DP Route
 Error. This reverts _call and DP Route Error code, and it sets @robust
 retries to only 1, since UIT+ should repair that connection immediately.
Also increased some timeouts from 20s to 30s, since Narwhal is a tad slow.

* Remove unused variables in exception handlers